feat: add ability to discover rules not in a set (#3284)
* perf: switch order of filtering to improve performance

It is faster to filter by group first, then by set, and lastly by query.
This lowers the number of rules that need to be filtered by more
expensive operations.

* feat: add find method to RectorFinder

Instead of having to call findCore and findCommunity separately and
merge them together, this adds a single find method that returns all
the rules.

* feat: add ability to discover rules not in a set